PrivateContract manticore test
Simple test to attempt to gain coverage over PrivateContract.

Currently, due to the alpha-stage of Manticore, some features need
to be implemented for this test to function properly.

A KeyError bug is causing the test to crash, but the test does
function somewhat. Will open an issue w/ ToB for the KeyError bug.